backend/requirements.txt
fastapi==0.116.0
langchain==0.3.26
langchain_core==0.3.68
langchain_google_genai==2.1.7
langchain_openai==0.3.27
langchain_pinecone==0.2.8
langgraph==0.5.2
langmem==0.0.28
pydantic==2.11.7
python-dotenv==1.1.1
typing_extensions==4.14.1
